refactor(adapters): consolidate IMetricsAdapter/ILogsAdapter into adapters
Canonical IMetricsAdapter / ILogsAdapter / IChangesAdapter /
IWebSearchAdapter now live in adapters/src/interfaces.ts. Concrete
classes (PrometheusMetricsAdapter, LokiLogsAdapter,
DuckDuckGoSearchAdapter) explicitly implement them — drops the
'avoid circular dep' duck-typing comments. agent-core re-exports
the same types so existing internal imports keep working.
Removes the unused PrometheusAdapter (DataAdapter shape) duplicate.
