+++ packages/adapters/src/loki/index.ts
@@ -1,6 +1,1 @@
 export { LokiLogsAdapter } from './logs-adapter.js';
-export type {
-  LogEntry,
-  LogsQueryInput,
-  LogsQueryResult,
-} from './logs-adapter.js';
