+++ packages/adapters/src/prometheus/index.ts
@@ -1,5 +1,4 @@
 export * from './types.js';
 export * from './translator.js';
-export * from './PrometheusAdapter.js';
 export * from './client.js';
 export * from './metrics-adapter.js';
