+++ packages/adapters/src/web-search/index.ts
@@ -1,2 +1,1 @@
 export { DuckDuckGoSearchAdapter } from './duckduckgo-adapter.js';
-export type { WebSearchResult } from './duckduckgo-adapter.js';
